#53 [FIX] 유저 칩 이름 같을 때만 '나'로 보이게 로직 수정
